can use condition on sass now
